Balloon Yak: new Zot monster (regret-index)
A chaos-themed yak variant found in the realm of Zot. On death bursts with a wind blast and a spread of chaotic clouds. During life, sucks the player towards it. Spawns in rather smaller packs than more mundane yaks as the clouds could get a bit ridiculous.

Made by a street performer wizard.