Yaktaur Glyphcaster: new monster
A new addition to Yaktaur bands, the glyphcaster has Sigil of Binding
in its spellbook. Stat-wise it is in the middle of Captains and plain
Yaktaurs.

Sigil works well whether or not the taurs end up binding themselves,
since they can still hit the player with ranged attacks. The player
has a harder time either approaching or running away from the threat,
and the yaktaurs will be hasted after being bound.